// src/enums/microbio.enums.ts
export enum EMedium {
  BLOOD_AGAR = "Кровяной агар",
  SABURO = "Агар сабуро",
}

export enum ESex {
  MALE = "Мужской",
  FEMALE = "Женский",
}

export enum ESusceptibility {
  S = "S",
  I = "I",
  R = "R",
  ATU = "ATU",
  NA = "NA",
}

export enum EExpertSystem {
  EUCST25 = "EUCST25",
}